+++ lib/models/week_forecast.dart
@@ -2,7 +2,7 @@ class WeekForecast {
   const WeekForecast(this.label, this.minTemp, this.maxTemp, this.weatherType);
 
   final String label;
-  final int minTemp;
-  final int maxTemp;
+  final double minTemp;
+  final double maxTemp;
   final String weatherType;
 }
